Added CommandCharacter element to WorldServer.ini file to tell the core what character to use for system commands in the chat.
 If you put a . in the CommandCharacter element, then you would do system commands like .info to get info on a targeted npc, if you put an ! in the CommandCharacter you would do the command like !info.

The Warden Bot will whisper the command response as usual, but no more whisper warden for commands.

Changed the Help command to list each command on a separate line, the way it was, would not fit in the chat area, and you could not see the list after it showed.

Added default AI to npcflood command.

Fixed the Chat command response.

Fixed a couple of typos in comments.